fix(workspace): clear rebrand follow-up debt
