+++ packages/unifia/src/global/index.ts
@@ -1,30 +1,9 @@
 import fs from "node:fs/promises"
-import { xdgData, xdgCache, xdgConfig, xdgState } from "xdg-basedir"
 import path from "node:path"
-import os from "node:os"
 import { Filesystem } from "../util/filesystem"
+import { Global } from "./path"
 
-const app = "unifia"
-
-const data = path.join(xdgData!, app)
-const cache = path.join(xdgCache!, app)
-const config = path.join(xdgConfig!, app)
-const state = path.join(xdgState!, app)
-
-export namespace Global {
-  export const Path = {
-    // Allow override via UNIFIA_TEST_HOME for test isolation
-    get home() {
-      return process.env.UNIFIA_TEST_HOME || os.homedir()
-    },
-    data,
-    bin: path.join(cache, "bin"),
-    log: path.join(data, "log"),
-    cache,
-    config,
-    state,
-  }
-}
+export { Global } from "./path"
 
 await Promise.all([
   fs.mkdir(Global.Path.data, { recursive: true }),
